refactor: reorganize packages out of store/ into proper locations
